+++ src/c65_startup.hpp
@@ -17,7 +17,9 @@ void setControlMode(bool);  // In CTRL menu
 
 enum startupState_t
 {
-    StartupIsInHomePosition,
+    StartupAskIfRAHomingShouldRun,
+    StartupAskIfDECHomingShouldRun, //DEC
+    StartupAskIfIsInHomePosition,
     StartupSetRoll,
     StartupWaitForRollCompletion,
     StartupRollConfirmed,
@@ -33,8 +35,21 @@ enum startupState_t
         #define NO     2
         #define CANCEL 3
 
-startupState_t startupState = StartupIsInHomePosition;
-int isInHomePosition        = NO;
+        #if USE_HALL_SENSOR_RA_AUTOHOME == 1
+startupState_t startupState = StartupAskIfRAHomingShouldRun;
+        #else
+startupState_t startupState = StartupAskIfIsInHomePosition;
+        #endif
+
+//DEC-
+        #if USE_HALL_SENSOR_DEC_AUTOHOME == 1
+startupState_t startupState = StartupAskIfDECHomingShouldRun;
+        #else
+startupState_t startupState = StartupAskIfIsInHomePosition;
+        #endif
+//-DEC
+
+int isInHomePosition = NO;
 
 void startupIsCompleted()
 {
@@ -44,7 +59,7 @@ void startupIsCompleted()
     inStartup      = false;
     okToUpdateMenu = true;
 
-    mount.startSlewing(TRACKING);
+////    mount.startSlewing(TRACKING);
 
     // Start on the RA menu
     lcdMenu.setActive(RA_Menu);
@@ -57,7 +72,9 @@ bool processStartupKeys()
     bool waitForRelease = false;
     switch (startupState)
     {
-        case StartupIsInHomePosition:
+        case StartupAskIfRAHomingShouldRun: // RA
+        case StartupAskIfDECHomingShouldRun: // DEC
+        case StartupAskIfIsInHomePosition:
             {
                 if (lcdButtons.keyChanged(&key))
                 {
@@ -155,7 +172,7 @@ bool processStartupKeys()
                 isInHomePosition = YES;
 
                 // Ask again to confirm
-                startupState = StartupIsInHomePosition;
+                startupState = StartupAskIfIsInHomePosition;
             }
             break;
 
@@ -170,8 +187,10 @@ void printStartupMenu()
 {
     switch (startupState)
     {
-        case StartupIsInHomePosition:
-            {
+        case StartupAskIfRAHomingShouldRun: // RA
+        case StartupAskIfDECHomingShouldRun: // DEC
+
+        case StartupAskIfIsInHomePosition:            {
                 //              0123456789012345
                 String choices(" Yes  No  Cancl ");
                 if (isInHomePosition == YES)